HU-03 – Seguimiento de Pedidos

+++ jama/src/main/java/com/cibertec/jama/repositories/pedido/PedidoItemRepository.java
@@ -2,7 +2,6 @@ package com.cibertec.jama.repositories.pedido;
 
 import com.cibertec.jama.entities.pedido.PedidoItem;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface PedidoItemRepository extends JpaRepository<PedidoItem, Long> {
+public interface PedidoItemRepository extends JpaRepository<PedidoItem, Integer> {
 }
